work around openSUSE %python_subpackage macro
Add the macro name to the list of names marking the end of the preamble,
as we're only interested in that. This is necessary because it gets
expanded to %package directives in place and rpmspectool attempts to
print out the preamble in the %prep step of an intermediate spec file,
which will result in duplicate %package directives.

Signed-off-by: Nils Philippsen <nils@tiptoe.de>